1st tier of Term Project
1st tier of the term project, accesses PubMed database using BioPython library to directly interface with the PubMed API. The program scrapes the articles from each of the four assigned categories, then compiles them into a Pandas DataFrame. This DataFrame is then converted to a CSV file, allowing for storage of all the abstracts without accessing the the PubMed database each time the code is executed. 

In the PubMed database, occasionally articles that are tagged with having abstracts instead are empty on that field. This causes the program to throw an error, which would crash the program. In order to compensate for this, a try except was utilized to throw out any files without an abstract. When this occurs, "whoopsies!" will be displayed on the console.