Second and third tier of Term Project
Second and third tiers of the term project, Obtains the abstract text data from the 1st tier of the term project, uses Pandas to read a CSV file to obtain abstract data, then normalizes the text data through case normalization, removal of punctuation, tokenizing, eliminating stop words, stemming the tokens, and lemmatizing the tokens, then creates a Bag of Words (BoW) representation for each abstract and stores the representation for the words and count separate lists, which is then converted into a separate list of lists, then bound together in a dictionary. The classifications of each abstract are also made into a list at this point.

The dictionary of the BoW words and count is then made into a Pandas DataFrame, then searched through to replace any NaN values with 0 for proper information feeding into the model. Then the BoW DataFrame and the list of classifications are sent into the train_test_split method from SKlearn, are shuffled 42 times, then split with a 70% train to 30% test ratio of the data, which is what has been found to be the most effective in terms of accuracy. The data is then fit using the Multinomial Naive Bayes, One versus Rest and One versus One Support Vector Machine using Linear Support Vector Classifier, and Logistical Regression from Sklearn. Then using the confusion_matrix, accuracy_score, precision_score, recall_score, and f1_score methods from SKlearn, the performance of each model is examined using the 'macro' parameter since we are working with a large volume of data. The matrix and scores are then printed for evaluation.

The preprocessing of all data to be fed into each algorithm is the same since they are all from the SKlearn package, leading to a simple conversion process from algorithm to algorithm.